moved Department enum to enums.dart file
